natural_language_query_engine/requirements.txt
gradio==4.43.0
gradio_client==1.3.0
python-dotenv==1.0.1
couchbase==4.3.1
langchain==0.2.16
langchain-community==0.2.16
openai==1.43.1
streamlit==1.38.0
streamlit-audiorecorder==0.0.5